Simplify dependency requirements for tests
I had problems when trying to integrate the #68 email notifications code in
to the code base as the existing tests pulled in every artifact (via
GeodesyServiceTestConfig via GeodesyServiceConfig via a
"use base-package "au.gov.ga.geodesy.domain.service" to scan for components).
The trouble with this is that every class, whether wanted or not (and if
relevant to the test or not) are included and wired up. An example of why
this was an issue is - one test had no explicit reference to the NodeService,
but because it was wired in automatically, this service had an effect.
It turned out this side-effect was desired but I could see how easily a
side-effect may cause undesired effects.

So I removed the \@ContextConfiguration that included the GeodesySupportConfig.class
and replaced each test with explicit references to the classes that they used.

This commit also removes the use of AspectJ for tests. This was only required in
one place - SiteLogSopacReader and its super-class SiteLogReader.

All test pass and can now be run in Eclipse :)  (now that the AspectJ weaving has been removed).