Fix issue with TimePrimitivePropertyType in Dozer mapper
TimePrimitivePropertyType can be resolved as a number of other
classes (via delegation) - we (so far) only use TimeInstantType for
a single date and TimePeriodType for a period (begin and end - the
end being optional to leave the period unidentified).

In this log, TimePrimitivePropertyType will be called 'TimePrimitive;'
TimeInstantType 'Instant'; and TimePeriodType 'Period'

This fix enforces the use of this type (via TimePrimitivePropertyTypeUtils)
to:
* A TimePrimitive created with a single value returns
a TimePrimitive with an Instant (unless null)
* A TimePrimitive created with a two values returns
a TimePrimitive with an Period (unless null)
* If the begin or end is null then the Period's BeginPosition
or EndPosition is null
* A null for all the arguments returns a TimePrimitive without an Instant or
Period payload.